fix broken storeutils
