+++ google/storeutils/const.go
