+++ google/storeutils/utils.go
@@ -1,0 +1,18 @@
+package storeutils
+
+import "github.com/lytics/cloudstorage"
+
+// GetAndOpen is a convenience method that combines Store.Get() and Object.Open() into
+// a single call.
+func GetAndOpen(s cloudstorage.Store, o string, level cloudstorage.AccessLevel) (cloudstorage.Object, error) {
+	obj, err := s.Get(o)
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = obj.Open(level)
+	if err != nil {
+		return nil, err
+	}
+	return obj, nil
+}
